+++ cordapp/src/main/kotlin/net/corda/plugins/Info.kt
@@ -1,3 +1,3 @@
 package net.corda.plugins
 
-open class Info(var name: String? = null, var version: String? = null, var vendor: String? = null, var targetVersion: Int? = null, var minPlatformVersion: Int? = null)
+open class Info(var name: String? = null, var version: String? = null, var vendor: String? = null, var targetPlatformVersion: Int? = null, var minimumPlatformVersion: Int? = null)